refactor(prefilled wallet): move to its own module
move the prefilled wallet into its own module since more than one api
module depends on it
